+++ src/App.js
@@ -2,7 +2,6 @@ import React, { useEffect } from "react";
 import { Switch, Route } from "react-router-dom";
 
 import Navigation from "./components/Navigation";
-import Loading from "./components/Loading";
 import MessageBox from "./components/MessageBox";
 import Home from "./pages/Home";
 import SignUp from "./pages/SignUp";
@@ -16,13 +15,11 @@ import CreateEvent from "./pages/CreateEvent";
 import Support from "./pages/Support";
 import NotFound from "./pages/NotFound";
 
-import { useDispatch, useSelector } from "react-redux";
-import { selectAppLoading } from "./store/appState/selectors";
+import { useDispatch } from "react-redux";
 import { getUserWithStoredToken } from "./store/user/actions";
 
 function App() {
   const dispatch = useDispatch();
-  const isLoading = useSelector(selectAppLoading);
 
   useEffect(() => {
     dispatch(getUserWithStoredToken());
