+++ src/components/Navigation/LoggedOut.js
@@ -3,6 +3,7 @@ import { Link as RouterLink } from "react-router-dom";
 import { Button } from "@material-ui/core";
 
 export default function LoggedOut() {
+  const ref = React.createRef();
   return (
     <>
       <Button
@@ -10,6 +11,7 @@ export default function LoggedOut() {
         variant="outlined"
         component={RouterLink}
         to="/login"
+        ref={ref}
       >
         LOGIN
       </Button>
