+++ src/components/Navigation/index.js
@@ -1,23 +1,20 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import { Link as RouterLink, NavLink } from "react-router-dom";
+import { Link as RouterLink } from "react-router-dom";
 
 import AppBar from "@material-ui/core/AppBar";
 import Toolbar from "@material-ui/core/Toolbar";
 import Typography from "@material-ui/core/Typography";
-import Tabs from "@material-ui/core/Tabs";
-import Tab from "@material-ui/core/Tab";
 import Link from "@material-ui/core/Link";
 import { makeStyles } from "@material-ui/core/styles";
-import GroupIcon from "@material-ui/icons/Group";
-import EventNoteIcon from "@material-ui/icons/EventNote";
 import EmojiEventsIcon from "@material-ui/icons/EmojiEvents";
+import { IconButton } from "@material-ui/core";
 
 import LoggedIn from "./LoggedIn";
 import LoggedOut from "./LoggedOut";
 
 import { selectToken } from "../../store/user/selectors";
-import { IconButton } from "@material-ui/core";
+import { selectUser } from "../../store/user/selectors";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -29,9 +26,6 @@ const useStyles = makeStyles((theme) => ({
   title: {
     flexGrow: 1,
   },
-  tabs: {
-    flexGrow: 1,
-  },
   link: {
     margin: theme.spacing(1, 1.5),
   },
@@ -40,14 +34,9 @@ const useStyles = makeStyles((theme) => ({
 export default function Navigation() {
   const token = useSelector(selectToken);
   const loginLogoutControls = token ? <LoggedIn /> : <LoggedOut />;
-
   const classes = useStyles();
-
-  const [value, setValue] = React.useState(0);
-
-  const handleChange = (event, newValue) => {
-    setValue(newValue);
-  };
+  const user = useSelector(selectUser);
+  const ref = React.createRef();
 
   return (
     <div className={classes.root}>
@@ -70,11 +59,24 @@ export default function Navigation() {
             component={RouterLink}
             variant="button"
             color="inherit"
+            ref={ref}
             className={classes.link}
             to="/events"
           >
             EVENTS
           </Link>
+          {token ? (
+            <Link
+              variant="button"
+              color="inherit"
+              ref={ref}
+              className={classes.link}
+              component={RouterLink}
+              to={`/teammate/${user.id}`}
+            >
+              {user.name}
+            </Link>
+          ) : null}
           {loginLogoutControls}
         </Toolbar>
       </AppBar>
