+++ src/pages/EventDetails/EditEventForm.js
@@ -1,6 +1,5 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { useHistory } from "react-router-dom";
 
 import {
   Container,
